Introduce low-level API module
Prior to this commit, the consumer-facing `Extacct` module was
responsible for directing most of the logic necessary to create Intacct
messages and sending them to the Intacct endpoint.  This commit
introduces a new API module which handles all of the low-level API
interfacing necessary with Intacct.  The externally-facing API will no
longer bear any direct responsibility for managing the behaviors
necessary to communicate with Intacct.  Instead, the new API module
(with its corresponding helper modules in the namespace) will handle all
message building and communication with Intacct.