GetList GenServer Implementation
Prior to this commit, users had no mechanism by which additional records
could be retrieved via the `get_list` API method.

This commit introduces several changes to the behavior of `get_list` as
well as a minor refactor:

1.  `get_list` now follows the same approach as the `read_report`
function.  Namely, API consumers must now provide a process that
conforms to the behaviours specified in the `GetListHandler` module.
This will allow for the processing of records as they are returned from
the Intacct API Endpoint.

2.  The `read`, `readByName`, and `readByQuery` methods have been removed
from the `RequestHandler` implemenatation.  The methods were only
available internally and currently had no hooks to external
client-facing calls.