Gateway Response Refactor
Prior to this commit, consumers of the Extacct response messages had
inconsistent treatment of record metadata.  The record metadata itself,
given the current `GenServer` implementation in place for `read_report`
and `get_list` provided little value.

This commit introduces an internal struct for use by the low-level API
and `RequestWorker` for handling the endpoint responses.  With this
change, the GatewayResponse will hold the content of the endpoint
response, and a separate struct containing the record metadata.

With these changes, it will be possible to correlate response messages
from Intacct together using the `control_id` contained in the metadata.