Add read_by_query record handling capability
Prior to this commit, the `read_by_query` function calls completely
ignored the record metadata returned by Intacct, obfuscating the ability
for users to understand the number of records read and remaining for a
given query.

This commit resolves this problem two-fold; first, by surfacing the
record metadata as part of the response processing, and secondly, by
converting to the `RequestWorker` model for queries.  Clients will now
need to provide a handler to the `read_by_query` function for handling
responses from the Intacct endpoint.  With this change, it will be
possible to receive and process all records returned by the query.