Provide state on callbacks
Prior to this commit, there was no consistency to the state returned to
implementers of th callback handlers.  This commit updates the handlers
to consistently provide the currently known state as a response to the
`GenServer` upon execution.  This will ensure that clients who provide a
handler will receive a consistently-managed state, as opposed to an
aribtrarly-assigned state as determined by the Extacct library.